+++ lib/easy_localization.dart
@@ -4,4 +4,5 @@ export 'package:easy_localization/src/easy_localization_app.dart';
 export 'package:easy_localization/src/asset_loader.dart';
 export 'package:easy_localization/src/public.dart';
 export 'package:easy_localization/src/public_ext.dart';
+export 'package:easy_localization/src/bloc/easy_localization_bloc.dart';
 export 'package:intl/intl.dart';