+++ lib/src/easy_localization_app.dart
@@ -1,111 +1,196 @@
+import 'dart:async';
 import 'dart:developer';
 
 import 'package:flutter/material.dart';
+import 'package:intl/intl_standalone.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+import 'package:easy_localization/src/widgets.dart';
 
 import 'asset_loader.dart';
 import 'localization.dart';
+import 'translations.dart';
+
+import 'bloc/easy_localization_bloc.dart';
 
 class EasyLocalization extends StatefulWidget {
+  final Widget child;
   final List<Locale> supportedLocales;
+  final Locale fallbackLocale;
   final bool useOnlyLangCode;
   final String path;
   final AssetLoader assetLoader;
-  final Widget child;
   final bool saveLocale;
-  final EasyLocalizationDelegate delegate;
-
+  final Color preloaderColor;
   EasyLocalization({
     Key key,
+    @required this.child,
     @required this.supportedLocales,
     @required this.path,
+    this.fallbackLocale,
     this.useOnlyLangCode = false,
     this.assetLoader = const RootBundleAssetLoader(),
     this.saveLocale = true,
-    this.child,
-  })  : delegate = EasyLocalizationDelegate(
-          path: path,
-          supportedLocales: supportedLocales,
-          useOnlyLangCode: useOnlyLangCode,
-          assetLoader: assetLoader,
-        ),
-        super(key: key);
+    this.preloaderColor = Colors.white,
+  })  : assert(child != null),
+        assert(supportedLocales != null && supportedLocales.isNotEmpty),
+        assert(path != null && path.isNotEmpty),
+        super(key: key) {
+    log("EasyLocalization");
+  }
 
-  static EasyLocalizationProvider of(BuildContext context) =>
-      EasyLocalizationProvider.of(context);
+  _EasyLocalizationState createState() => _EasyLocalizationState();
 
-  @override
-  _EasyLocalizationState createState() => _EasyLocalizationState(delegate);
+  static _EasyLocalizationProvider of(BuildContext context) =>
+      _EasyLocalizationProvider.of(context);
 }
 
 class _EasyLocalizationState extends State<EasyLocalization> {
+  ///Init EasyLocalizationBloc
+  final EasyLocalizationBloc bloc = EasyLocalizationBloc();
+  _EasyLocalizationDelegate delegate;
   Locale locale;
-  final EasyLocalizationDelegate delegate;
 
-  _EasyLocalizationState(this.delegate);
-
-  _onLocaleChanged(Locale locale) {
-    setState(() {
-      this.locale = locale;
-    });
+  @override
+  void dispose() {
+    bloc.dispose();
+    super.dispose();
   }
 
   @override
   void initState() {
-    if (widget.saveLocale) loadSavedLocale();
-    delegate.onLocaleChanged = (locale) {
-      if (this.locale == null)
-        setState(() {
-          this.locale = locale;
-        });
-    };
+    log("initState");
+    _init();
     super.initState();
   }
 
-  loadSavedLocale() async {
-    SharedPreferences _preferences = await SharedPreferences.getInstance();
-    var _strLocale = _preferences.getString('locale');
-    if (_strLocale != null) {
-      log('easy localization: Locale loaded from shared preferences ${_strLocale}');
-      setState(() {
-        locale = _localeFromString(_strLocale);
-      });
+  @override
+  void reassemble() {
+    bloc.reassemble();
+    super.reassemble();
+  }
+
+  _init() async {
+    Locale _savedLocale;
+    Locale _osLocale;
+    if (widget.saveLocale) _savedLocale = await loadSavedLocale();
+    // Get Device Locale
+    _osLocale = await _getDeviceLocale();
+    // If saved locale then get
+    if (_savedLocale != null && widget.saveLocale) {
+      log('easy localization: Saved locale loaded ${_savedLocale.toString()}');
+      locale = _savedLocale;
+    } else {
+      locale = widget.supportedLocales.firstWhere(
+          (locale) => _checkInitLocale(locale, _osLocale),
+          orElse: () => _getFallbackLocale(
+              widget.supportedLocales, widget.fallbackLocale));
     }
-    // TODO reload delegate, set on Material Widget
+
+    bloc.onChange(Resource(
+        locale: locale,
+        assetLoader: widget.assetLoader,
+        path: widget.path,
+        useOnlyLangCode: widget.useOnlyLangCode));
+  }
+
+  bool _checkInitLocale(Locale locale, Locale _osLocale) {
+    // If suported locale not contain countryCode then check only languageCode
+    if (locale.countryCode == null) {
+      return (locale.languageCode == _osLocale.languageCode);
+    } else {
+      return (locale == _osLocale);
+    }
+  }
+
+  //Get fallback Locale
+  Locale _getFallbackLocale(
+      List<Locale> supportedLocales, Locale fallbackLocale) {
+    //If fallbackLocale not set then return first from supportedLocales
+    if (fallbackLocale != null) {
+      return fallbackLocale;
+    } else {
+      return supportedLocales.first;
+    }
+  }
+
+  // Get Device Locale
+  Future<Locale> _getDeviceLocale() async {
+    final String _deviceLocale = await findSystemLocale();
+    log('easy localization: Device locale $_deviceLocale');
+    return _localeFromString(_deviceLocale);
+  }
+
+  Future<Locale> loadSavedLocale() async {
+    SharedPreferences _preferences = await SharedPreferences.getInstance();
+    final _strLocale = _preferences.getString('locale');
+    final locale = _strLocale != null ? _localeFromString(_strLocale) : null;
+    
+    return locale;
   }
 
   @override
   Widget build(BuildContext context) {
-    return EasyLocalizationProvider(
-      widget,
-      this.locale,
-      child: widget.child,
-      onLocaleChanged: _onLocaleChanged,
-      delegate: delegate,
+    Widget returnWidget;
+    log("easy localization: Build");
+    return Container(
+      color: widget.preloaderColor,
+      child: StreamBuilder<Resource>(
+          stream: bloc.outStream,
+          builder: (context, snapshot) {
+            if (snapshot.connectionState == ConnectionState.waiting &&
+                !snapshot.hasData &&
+                !snapshot.hasError) {
+              returnWidget = EmptyPreloaderWidget();
+            } else if (snapshot.hasData && !snapshot.hasError) {
+              returnWidget = _EasyLocalizationProvider(
+                widget,
+                snapshot.data.locale,
+                bloc: bloc,
+                delegate: _EasyLocalizationDelegate(
+                    translations: snapshot.data.translations,
+                    supportedLocales: widget.supportedLocales),
+              );
+            } else if (snapshot.hasError) {
+              returnWidget = FutureErrorWidget(msg: snapshot.error);
+            }
+            return returnWidget;
+          }),
     );
   }
 }
 
-class EasyLocalizationProvider extends InheritedWidget {
+class _EasyLocalizationProvider extends InheritedWidget {
   final EasyLocalization parent;
   final Locale _locale;
-  final ValueChanged<Locale> onLocaleChanged;
+  final EasyLocalizationBloc bloc;
+  final _EasyLocalizationDelegate delegate;
 
   List<Locale> get supportedLocales => parent.supportedLocales;
 
-  final EasyLocalizationDelegate delegate;
+  // _EasyLocalizationDelegate get delegate => parent.delegate;
 
-  EasyLocalizationProvider(this.parent, this._locale,
-      {Key key, Widget child, this.onLocaleChanged, this.delegate})
-      : super(key: key, child: child);
+  _EasyLocalizationProvider(this.parent, this._locale,
+      {Key key, this.bloc, this.delegate})
+      : super(key: key, child: parent.child) {
+    log('easy localization: Init provider');
+  }
 
   Locale get locale => _locale;
+  Locale get fallbackLocale => parent.fallbackLocale;
 
   set locale(Locale locale) {
-    assert(parent.supportedLocales.contains(locale));
-    if (parent.saveLocale) _saveLocale(locale);
-    log('easy localization: Locale set ${locale.toString()}');
-    onLocaleChanged(locale);
+    // Check old locale
+    if (locale != _locale) {
+      assert(parent.supportedLocales.contains(locale));
+      if (parent.saveLocale) _saveLocale(locale);
+      log('easy localization: Locale set ${locale.toString()}');
+
+      bloc.onChange(Resource(
+          locale: locale,
+          path: parent.path,
+          assetLoader: parent.assetLoader,
+          useOnlyLangCode: parent.useOnlyLangCode));
+    }
   }
 
   _saveLocale(Locale locale) async {
@@ -121,52 +206,38 @@ class EasyLocalizationProvider extends InheritedWidget {
   }
 
   @override
-  bool updateShouldNotify(EasyLocalizationProvider oldWidget) {
+  bool updateShouldNotify(_EasyLocalizationProvider oldWidget) {
     return oldWidget._locale != this._locale;
   }
 
-  static EasyLocalizationProvider of(BuildContext context) =>
-      context.dependOnInheritedWidgetOfExactType<EasyLocalizationProvider>();
+  static _EasyLocalizationProvider of(BuildContext context) =>
+      context.dependOnInheritedWidgetOfExactType<_EasyLocalizationProvider>();
 }
 
-class EasyLocalizationDelegate extends LocalizationsDelegate<Localization> {
-  final String path;
-  final AssetLoader assetLoader;
+class _EasyLocalizationDelegate extends LocalizationsDelegate<Localization> {
   final List<Locale> supportedLocales;
-  ValueChanged<Locale> onLocaleChanged;
+  final Translations translations;
 
   ///  * use only the lang code to generate i18n file path like en.json or ar.json
-  final bool useOnlyLangCode;
+  // final bool useOnlyLangCode;
 
-  Locale loadedLocale;
-
-  EasyLocalizationDelegate({
-    @required this.path,
-    @required this.supportedLocales,
-    this.useOnlyLangCode = false,
-    this.assetLoader,
-  });
+  _EasyLocalizationDelegate({this.translations, this.supportedLocales}) {
+    log('easy localization: Init Localization Delegate');
+    Localization.instance.translations = translations;
+  }
 
   @override
   bool isSupported(Locale locale) => supportedLocales.contains(locale);
 
   @override
-  Future<Localization> load(Locale value) async {
-    loadedLocale = value;
-    await Localization.load(
-      value,
-      path: path,
-      useOnlyLangCode: useOnlyLangCode,
-      assetLoader: assetLoader,
-    );
-    onLocaleChanged(value);
-    return Localization.instance;
+  Future<Localization> load(Locale value) {
+    log('easy localization: Load Localization Delegate');
+    Localization.load(value, translations: translations);
+    return Future.value(Localization.instance);
   }
 
   @override
-  bool shouldReload(EasyLocalizationDelegate old) {
-    return loadedLocale != old.loadedLocale;
-  }
+  bool shouldReload(LocalizationsDelegate<Localization> old) => false;
 }
 
 Locale _localeFromString(String val) {