+++ test/easy_localization_widget_test.dart
@@ -6,6 +6,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'utils/test_asset_loaders.dart';
 
+import 'package:flutter_localizations/flutter_localizations.dart';
+
 BuildContext _context;
 
 class MyApp extends StatelessWidget {
@@ -15,6 +17,8 @@ class MyApp extends StatelessWidget {
       locale: EasyLocalization.of(context).locale,
       supportedLocales: EasyLocalization.of(context).supportedLocales,
       localizationsDelegates: [
+        GlobalMaterialLocalizations.delegate,
+        GlobalWidgetsLocalizations.delegate,
         EasyLocalization.of(context).delegate,
       ],
       home: MyWidget(),
@@ -55,12 +59,11 @@ void main() {
         await tester.pumpAndSettle();
 
         expect(Localization.of(_context), isInstanceOf<Localization>());
+        expect(Localization.instance, isInstanceOf<Localization>());
+        expect(Localization.instance, Localization.of(_context));
         expect(EasyLocalization.of(_context).supportedLocales,
             [Locale("en", "US")]);
         expect(EasyLocalization.of(_context).locale, Locale("en", "US"));
-//        expect(Intl.defaultLocale, Locale("en", "US").toString());
-//        expect(Intl.defaultLocale,
-//            EasyLocalization.of(_context).locale.toString());
 
         final trFinder = find.text('test');
         expect(trFinder, findsOneWidget);
@@ -71,6 +74,9 @@ void main() {
         expect(plural("day", 1, context: _context), "1 day");
         expect(plural("day", 2, context: _context), "2 days");
         expect(plural("day", 3, context: _context), "3 other days");
+
+        expect("test".tr(), "test");
+        expect("day".plural(1), "1 day");
       });
     },
   );
@@ -92,9 +98,6 @@ void main() {
         expect(EasyLocalization.of(_context).supportedLocales,
             [Locale("en", "US")]);
         expect(EasyLocalization.of(_context).locale, Locale("en", "US"));
-//        expect(Intl.defaultLocale, Locale("en", "US").toString());
-//        expect(Intl.defaultLocale,
-//            EasyLocalization.of(_context).locale.toString());
 
         final trFinder = find.text('test');
         expect(trFinder, findsOneWidget);
@@ -124,9 +127,6 @@ void main() {
         expect(EasyLocalization.of(_context).supportedLocales,
             [Locale("en", "US")]);
         expect(EasyLocalization.of(_context).locale, Locale("en", "US"));
-//        expect(Intl.defaultLocale, Locale("en", "US").toString());
-//        expect(Intl.defaultLocale,
-//            EasyLocalization.of(_context).locale.toString());
 
         final trFinder = find.text('test');
         expect(trFinder, findsOneWidget);
@@ -140,7 +140,24 @@ void main() {
       });
     },
   );
-
+  testWidgets(
+    '[EasyLocalization with  Error path] test',
+    (WidgetTester tester) async {
+      await tester.runAsync(() async {
+        await tester.pumpWidget(EasyLocalization(
+          child: MyApp(),
+          path: "i18",
+          supportedLocales: [Locale("en", "US")],
+        ));
+        await tester.idle();
+        // The async delegator load will require build on the next frame. Thus, pump
+        await tester.pumpAndSettle();
+        final trFinder = find.text('Easy Localization:');
+        expect(trFinder, findsOneWidget);
+        await tester.pump();
+      });
+    },
+  );
   testWidgets(
     '[EasyLocalization] change loacle test',
     (WidgetTester tester) async {
@@ -148,21 +165,19 @@ void main() {
         await tester.pumpWidget(EasyLocalization(
           child: MyApp(),
           path: "i18n",
-          supportedLocales: [Locale("en", "US"), Locale("en")],
+          supportedLocales: [Locale("en", "US")],
         ));
         await tester.idle();
         // The async delegator load will require build on the next frame. Thus, pump
         await tester.pumpAndSettle();
 
         expect(EasyLocalization.of(_context).supportedLocales,
-            [Locale("en", "US"), Locale("en")]);
+            [Locale("en", "US")]);
         expect(EasyLocalization.of(_context).locale, Locale("en", "US"));
-//        expect(Intl.defaultLocale, Locale("en", "US").toString());
-//        expect(Intl.defaultLocale,
-//            EasyLocalization.of(_context).locale.toString());
 
         var l = Locale("en", "US");
         EasyLocalization.of(_context).locale = l;
+        await tester.pumpAndSettle();
         expect(EasyLocalization.of(_context).locale, Locale("en", "US"));
 
         final trFinder = find.text('test');
@@ -175,22 +190,13 @@ void main() {
         expect(plural("day", 2, context: _context), "2 days");
         expect(plural("day", 3, context: _context), "3 other days");
         expect(EasyLocalization.of(_context).locale, Locale("en", "US"));
-//        expect(Intl.defaultLocale, Locale("en", "US").toString());
 
         l = Locale("ar", "DZ");
-        expect(() => {EasyLocalization.of(_context).locale = l},
-            throwsAssertionError);
-
-        expect(EasyLocalization.of(_context).supportedLocales,
-            [Locale("en", "US"), Locale("en")]);
-        expect(EasyLocalization.of(_context).locale, Locale("en", "US"));
-
-        l = Locale("en");
-        EasyLocalization.of(_context).locale = l;
+        expect(() {
+          EasyLocalization.of(_context).locale = l;
+        }, throwsAssertionError);
         await tester.pumpAndSettle();
-        expect(EasyLocalization.of(_context).locale, l);
-
-        EasyLocalization.of(_context).locale = Locale("en", "US");
+        expect(EasyLocalization.of(_context).locale, Locale("en", "US"));
       });
     },
   );
@@ -225,20 +231,20 @@ void main() {
 
         var l = Locale("en", "US");
         EasyLocalization.of(_context).locale = l;
+        await tester.pumpAndSettle();
         expect(EasyLocalization.of(_context).locale, l);
-//        expect(Intl.defaultLocale, l.toString());
 
         l = Locale("ar", "DZ");
         EasyLocalization.of(_context).locale = l;
+        await tester.idle();
         await tester.pumpAndSettle();
         expect(EasyLocalization.of(_context).locale, l);
-//        expect(Intl.defaultLocale, l.toString());
 
         l = Locale("en", "US");
         EasyLocalization.of(_context).locale = l;
+        await tester.idle();
         await tester.pumpAndSettle();
         expect(EasyLocalization.of(_context).locale, l);
-//        expect(Intl.defaultLocale, l.toString());
 
         l = Locale("en", "UK");
         expect(() => {EasyLocalization.of(_context).locale = l},
@@ -246,6 +252,7 @@ void main() {
 
         l = Locale("ar", "DZ");
         EasyLocalization.of(_context).locale = l;
+        await tester.idle();
         await tester.pumpAndSettle();
         expect(EasyLocalization.of(_context).locale, l);
       });
@@ -273,9 +280,6 @@ void main() {
         expect(EasyLocalization.of(_context).supportedLocales,
             [Locale("en", "US"), Locale("ar", "DZ")]);
         expect(EasyLocalization.of(_context).locale, Locale("ar", "DZ"));
-//        expect(Intl.defaultLocale, Locale("ar", "DZ").toString());
-//        expect(Intl.defaultLocale,
-//            EasyLocalization.of(_context).locale.toString());
 
         var trFinder = find.text('اختبار');
         expect(trFinder, findsOneWidget);
@@ -316,9 +320,6 @@ void main() {
         expect(EasyLocalization.of(_context).supportedLocales,
             [Locale("en"), Locale("ar")]);
         expect(EasyLocalization.of(_context).locale, Locale("en"));
-//        expect(Intl.defaultLocale, Locale("en").toString());
-//        expect(Intl.defaultLocale,
-//            EasyLocalization.of(_context).locale.toString());
 
         var l = Locale("en");
         EasyLocalization.of(_context).locale = l;
@@ -349,10 +350,6 @@ void main() {
             [Locale("en"), Locale("ar")]);
         expect(EasyLocalization.of(_context).locale, Locale("en"));
 
-//        expect(Intl.defaultLocale, Locale("en").toString());
-//        expect(Intl.defaultLocale,
-//            EasyLocalization.of(_context).locale.toString());
-
         var l = Locale("en");
         EasyLocalization.of(_context).locale = l;
         expect(EasyLocalization.of(_context).locale, l);
@@ -370,6 +367,7 @@ void main() {
           saveLocale: false,
           useOnlyLangCode: true,
           supportedLocales: [Locale("ar")],
+          fallbackLocale:Locale("ar") ,
         ));
         await tester.idle();
         // The async delegator load will require build on the next frame. Thus, pump
@@ -377,10 +375,7 @@ void main() {
 
         expect(EasyLocalization.of(_context).supportedLocales, [Locale("ar")]);
         expect(EasyLocalization.of(_context).locale, Locale("ar"));
-
-//        expect(Intl.defaultLocale, Locale("ar").toString());
-//        expect(Intl.defaultLocale,
-//            EasyLocalization.of(_context).locale.toString());
+        expect(EasyLocalization.of(_context).fallbackLocale, Locale("ar"));
       });
     },
   );
@@ -405,10 +400,7 @@ void main() {
 
         expect(EasyLocalization.of(_context).supportedLocales, [Locale("ar")]);
         expect(EasyLocalization.of(_context).locale, Locale("ar"));
-
-//        expect(Intl.defaultLocale, Locale("ar").toString());
-//        expect(Intl.defaultLocale,
-//            EasyLocalization.of(_context).locale.toString());
+        expect(EasyLocalization.of(_context).fallbackLocale, null);
       });
     },
   );
@@ -438,10 +430,7 @@ void main() {
           expect(EasyLocalization.of(_context).supportedLocales,
               [Locale("en", "US"), Locale("ar", "DZ")]);
           expect(EasyLocalization.of(_context).locale, Locale("en", "US"));
-
-//          expect(Intl.defaultLocale, Locale("ar", "DZ").toString());
-//          expect(Intl.defaultLocale,
-//              EasyLocalization.of(_context).locale.toString());
+          expect(EasyLocalization.of(_context).fallbackLocale, null);
         });
       },
     );
@@ -476,10 +465,7 @@ void main() {
           expect(EasyLocalization.of(_context).supportedLocales,
               [Locale("en"), Locale("ar")]);
           expect(EasyLocalization.of(_context).locale, Locale("ar"));
-
-//          expect(Intl.defaultLocale, Locale("ar").toString());
-//          expect(Intl.defaultLocale,
-//              EasyLocalization.of(_context).locale.toString());
+          expect(EasyLocalization.of(_context).fallbackLocale, null);
         });
       },
     );
@@ -513,10 +499,7 @@ void main() {
           expect(EasyLocalization.of(_context).supportedLocales,
               [Locale("en", "US"), Locale("ar", "DZ")]);
           expect(EasyLocalization.of(_context).locale, Locale("ar", "DZ"));
-
-//          expect(Intl.defaultLocale, Locale("en", "US").toString());
-//          expect(Intl.defaultLocale,
-//              EasyLocalization.of(_context).locale.toString());
+          expect(EasyLocalization.of(_context).fallbackLocale, null);
         });
       },
     );
@@ -543,8 +526,7 @@ void main() {
               [Locale("en", "US"), Locale("ar", "DZ")]);
           expect(EasyLocalization.of(_context).locale, Locale("en", "US"));
 
-//          expect(Intl.defaultLocale, Locale("en", "US").toString());
-//          expect(Intl.defaultLocale,
+          EasyLocalization.of(_context).locale = Locale("en", "US");
         });
       },
     );