+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [2.1.0]
+
+- Added Error widget.
+- fixed many issues.
+- Based on Bloc.
+- optimized and clean code more stability
+
 ## [2.0.2]
 
 - fixed many issues